+++ src/OSPSuite.MinimalImplementations/ExceptionManager.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using OSPSuite.Utility.Exceptions;
+using OSPSuite.Utility.Extensions;
+
+namespace OSPSuite.MinimalImplementations
+{
+    public class ExceptionManager : ExceptionManagerBase
+    {
+        public override void LogException(Exception ex)
+        {
+            Console.WriteLine(ex.FullMessage());
+        }
+    }
+}
