+++ src/OSPSuite.MinimalImplementations/OSPSuite.MinimalImplementations.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="OSPSuite.Utility" Version="4.1.1.1" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\OSPSuite.Assets\OSPSuite.Assets.csproj" />
+		<ProjectReference Include="..\OSPSuite.Core\OSPSuite.Core.csproj" />
+		<ProjectReference Include="..\OSPSuite.Infrastructure.Autofac\OSPSuite.Infrastructure.Autofac.csproj" />
+		<ProjectReference Include="..\OSPSuite.Infrastructure.Import\OSPSuite.Infrastructure.Import.csproj" />
+		<ProjectReference Include="..\OSPSuite.Infrastructure.Serialization\OSPSuite.Infrastructure.Serialization.csproj" />
+		<ProjectReference Include="..\OSPSuite.Infrastructure\OSPSuite.Infrastructure.csproj" />
+	</ItemGroup>
+</Project>
