--- src/OSPSuite.R/MinimalImplementations/ProgressManager.cs
@@ -1,19 +1,0 @@
-﻿using OSPSuite.Utility.Events;
-
-namespace OSPSuite.R.MinimalImplementations
-{
-   public class ProgressManager : IProgressManager
-   {
-      private readonly IEventPublisher _eventPublisher;
-
-      public ProgressManager(IEventPublisher eventPublisher)
-      {
-         _eventPublisher = eventPublisher;
-      }
-
-      public IProgressUpdater Create()
-      {
-         return new ProgressUpdater(_eventPublisher);
-      }
-   }
-}
