+++ src/OSPSuite.R/OSPSuite.R.csproj
@@ -27,6 +27,7 @@
     <ProjectReference Include="..\OSPSuite.Infrastructure.Autofac\OSPSuite.Infrastructure.Autofac.csproj" />
     <ProjectReference Include="..\OSPSuite.Infrastructure.Import\OSPSuite.Infrastructure.Import.csproj" />
     <ProjectReference Include="..\OSPSuite.Infrastructure\OSPSuite.Infrastructure.csproj" />
+    <ProjectReference Include="..\OSPSuite.MinimalImplementations\OSPSuite.MinimalImplementations.csproj" />
   </ItemGroup>
 
   <ItemGroup>
